slope: planar dask backends declare float32 meta dtype (#2686) (#2694)
The planar dask paths passed meta=np.array(()) / meta=cupy.array((),
which default to float64, while the chunk functions return float32.
The lazy dask dtype (float64) disagreed with the computed dtype
(float32) and with the numpy/cupy backends. Declare float32 meta to
match, mirroring the geodesic dask paths.

Adds regression tests asserting lazy dtype == computed dtype == numpy
dtype (float32) across numpy, cupy, dask+numpy, and dask+cupy.

+++ xrspatial/slope.py
@@ -90,7 +90,7 @@ def _run_dask_numpy(data: da.Array,
     out = data.map_overlap(_func,
                            depth=(1, 1),
                            boundary=_boundary_to_dask(boundary),
-                           meta=np.array(()))
+                           meta=np.array((), dtype=np.float32))
     return out
 
 
@@ -106,7 +106,7 @@ def _run_dask_cupy(data: da.Array,
     out = data.map_overlap(_func,
                            depth=(1, 1),
                            boundary=_boundary_to_dask(boundary, is_cupy=True),
-                           meta=cupy.array(()))
+                           meta=cupy.array((), dtype=cupy.float32))
     return out
